requirements.max.txt
alabaster==0.7.6
astroid==1.3.8
Babel==2.0
blockscore==4.1.0
cmsplugin-filer==0.10.2
coverage==4.0a5
Django==1.8.4
django-admin-enhancer==1.0.0
django-angular==0.7.15
django-appconf==1.0.1
django-axes==1.3.9
django-braces==1.8.0
django-classy-tags==0.6.2
django-cms==3.1.2
django-compressor==1.5
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-extensions==1.5.2
django-filer==0.9.12
django-flat-theme==0.9.3
django-grappelli==2.7.1
django-jenkins==0.17.0
django-meta==0.3.1
django-meta-mixin==0.1.1
django-mptt==0.7.4
django-nested-admin==2.1.2
django-parler==1.5
django-polymorphic==0.7.1
django-sekizai==0.8.2
Django-Select2==4.3.1
django-seo-js==0.3.1
django-taggit==0.17.0
django-taggit-autosuggest==0.2.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==3.0
djangocms-admin-style==0.2.7
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-inherit==0.1
djangocms-link==1.6.2
djangocms-page-meta==0.5.5
djangocms-page-sitemap==0.2.0
djangocms-picture==0.1
-e git+https://github.com/inonit/djangocms-restapi.git@0c95623c694f526e75d1113278552492f4cb9c24#egg=djangocms_restapi-master
djangocms-snippet==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.6.0
djangocms-video==0.1
djangorestframework==3.2.2
djangorestframework-bulk==0.2.1
djangorestframework-recursive==0.1.1
djangular==0.2.7
docutils==0.12
drf-json-api==0.1.1
easy-thumbnails==2.2
gnureadline==6.3.3
html5lib==0.999
httplib2==0.9.1
ipython==3.1.0
Jinja2==2.8
logilab-common==1.0.2
MarkupSafe==0.23
pep8==1.6.2
Pillow==2.9.0
psycopg2==2.6
py-bcrypt==0.4
Pygments==2.0.2
pylint==1.4.3
pytz==2015.4
raven==5.1.1
requests==2.2.1
shadowsocks==2.8.2
simplejson==3.8.0
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
sqlparse==0.1.16
twilio==4.5.0
Unidecode==0.4.18
